fix(strict): close every -Wconversion/-Wshadow warning, STRICT now passes 19/19
Completes the follow-up cleanup flagged in the Phase 5 commit
(ba52f41) — the STRICT build gate (-DSTRICT=ON on host CMake)
was opt-in but did not pass on the unmodified codebase due to
twelve pre-existing narrowing conversions and a handful of other
diagnostics that -Werror surfaced. This commit fixes every one so
the STRICT pipeline now matches the rest of the green suite.

Scope
-----
Pure defensive edits; no behaviour change on the host build
(ctest 19/19 both with and without STRICT) and no ABI change for
the target wire protocols (CCSDS + AX.25 byte-for-byte identical
to Phase 5 — goldens still match).

Result
------
  cmake -B build-strict -S . -DSTRICT=ON && cmake --build build-strict
  ctest --test-dir build-strict
  -> 100% tests passed, 0 tests failed out of 19

  make cppcheck        -> gate clean, exit 0 (unchanged)
  make sanitizers      -> 19/19 ASAN+UBSAN green (unchanged)
  make test            -> 19/19 default ctest green (unchanged)

Per-file diffs
--------------

1. firmware/stm32/Core/Src/ccsds.c
   * CalculateCRC16: added (uint16_t) cast on the XOR result that
     feeds back into crc so the int-promoted XOR doesn't narrow
     silently into the uint16_t accumulator. 0xFFFF -> 0xFFFFU.
   * BuildPacket: removed the `(0 << 13) |` version-field OR (the
     redundant-OR that cppcheck flagged as badBitmaskCheck) and
     moved the CCSDS primary-header layout explanation into a
     code comment so the bit positions stay self-documenting.
     Rewrote the bitwise composition of first_word / sequence
     with explicit (uint16_t) + U-suffixed literals.
   * Serialize: every `(x >> 8) & 0xFF -> uint8_t` store now
     carries an explicit (uint8_t) cast with the 0xFFU mask.
     offset += N narrowings likewise cast to uint16_t.
   * Parse: matching (uint16_t) casts on the primary-header
     assembly; the 8-byte timestamp walk casts buffer[offset++]
     to uint64_t before OR-ing into the 64-bit accumulator.
   * ValidateCRC: (uint16_t) cast on the 2-byte assembly + on
     the (length - 2U) argument so CalculateCRC16 sees a clean
     uint16_t count.

2. firmware/stm32/Core/Src/comm.c
   * Three ring-buffer modulo updates (COMM_Receive,
     COMM_UART_RxCallback, COMM_ProcessRxBuffer) now cast the
     `(head + 1) % SIZE` result to uint16_t. SIZE is a uint32_t
     define so the addition promotes; the cast was missing.

3. firmware/stm32/Core/Src/error_handler.c
   * error_log_index ring-update cast to uint16_t, same pattern.

4. firmware/stm32/Core/Src/payload.c
   * radiation_data.index ring-update cast to uint8_t (the index
     is 60-modulo; both operands now U-suffixed).

5. firmware/stm32/Core/Src/telemetry.c
   * Telemetry_PackError: split the `strlen(message)` assignment
     into a size_t raw_len with explicit clamp-then-narrow to
     uint16_t. offset += msg_len likewise cast.

6. firmware/stm32/Drivers/SBM20/sbm20.c
   * SBM20_Tick1s: window_idx = (idx + 1) % SIZE -> uint8_t cast.

7. firmware/stm32/Drivers/VirtualUART/virtual_uart.c
   * VirtualUART_Send: send()'s ssize_t return now stored into
     ssize_t (was int); the comparison against len uses
     (ssize_t)len so a partial send on loopback is still treated
     as failure without the narrow-from-long warning.
   * VirtualUART_Recv: guards max_bytes < 0 up front; passes
     (size_t)max_bytes to recv; narrows the ssize_t result back
     to int only after the negative-branch handled.

8. firmware/tests/unity/unity.h
   * UNITY_BEGIN() switched from a comma expression with trailing
     0 to a do { ... } while(0) block so `-Werror=unused-value`
     does not trip on the (now meaningless) trailing constant.
     Every existing UNITY_BEGIN() call site is a statement, so
     the do-while is a drop-in replacement.

9. firmware/tests/test_comm_integration.c
   * Missing `#include <string.h>` — was getting implicit
     memcpy under STRICT's `-Werror=implicit-function-declaration`.

10. scripts/sitl_fw.c
    * Added `#define _POSIX_C_SOURCE 200809L` /
      `#define _DEFAULT_SOURCE 1` before <unistd.h> so glibc
      exposes usleep() under -std=c11. Pre-existing implicit-decl
      warning (noted in earlier build log) now resolved.

11. firmware/CMakeLists.txt
    * STRICT flag list extended with `-Wno-error=unused-function`
      and `-Wno-error=unused-const-variable`. Sensor drivers
      (BME280 compensators, SunSensor face_channels) hide real
      code behind `#ifndef SIMULATION_MODE`, so the static
      helpers are legitimately unused on host — demoting those
      two warnings to non-fatal preserves the STRICT gate's
      value (real narrowings fail CI) without false positives.

12. README.md + docs/quality/static_analysis.md
    * Status tables now reflect STRICT passing; deferred-item
      list in README drops the "STRICT -Wconversion cleanup"
      entry; static_analysis doc lists this commit as the
      resolution for the twelve narrowing locations it
      previously tracked.

Why unsigned literals / U-suffixed constants
--------------------------------------------
Under -Wconversion, an expression like `(x >> 8) & 0xFF` promotes
both operands to int, does the mask in int, and the store back
to uint8_t is the narrowing. Writing `0xFFU` forces unsigned
type-promotion so no sign bit is at stake; the explicit
(uint8_t) cast on the store documents the intentional narrow
and silences the warning. Same pattern everywhere.

Why moving to STRICT matters
----------------------------
On the STM32F446 `int` is 32 bits and the narrowings above are
safe at runtime — but the same pattern on an 8-bit or 16-bit
MCU would truncate silently. Making every narrow explicit in
the source now means any future port can stay STRICT-clean
without a repeat audit.

+++ firmware/stm32/Core/Src/comm.c
@@ -101,7 +101,7 @@ uint16_t COMM_Receive(CommChannel_t channel, uint8_t *buffer,
     uint16_t count = 0;
     while (uhf_rx_tail != uhf_rx_head && count < max_length) {
         buffer[count++] = uhf_rx_buffer[uhf_rx_tail];
-        uhf_rx_tail = (uhf_rx_tail + 1) % COMM_RX_BUFFER_SIZE;
+        uhf_rx_tail = (uint16_t)((uhf_rx_tail + 1U) % COMM_RX_BUFFER_SIZE);
     }
 
     if (count > 0) {
@@ -199,7 +199,7 @@ bool COMM_IsConnected(CommChannel_t channel) {
 
 void COMM_UART_RxCallback(CommChannel_t channel, uint8_t byte) {
     (void)channel;
-    uint16_t next = (uhf_rx_head + 1) % COMM_RX_BUFFER_SIZE;
+    uint16_t next = (uint16_t)((uhf_rx_head + 1U) % COMM_RX_BUFFER_SIZE);
     if (next != uhf_rx_tail) {
         uhf_rx_buffer[uhf_rx_head] = byte;
         uhf_rx_head = next;
@@ -213,7 +213,7 @@ void COMM_ProcessRxBuffer(void) {
      * COMM_Status_t so existing telemetry sees link-layer health. */
     while (uhf_rx_tail != uhf_rx_head) {
         uint8_t byte = uhf_rx_buffer[uhf_rx_tail];
-        uhf_rx_tail = (uhf_rx_tail + 1) % COMM_RX_BUFFER_SIZE;
+        uhf_rx_tail = (uint16_t)((uhf_rx_tail + 1U) % COMM_RX_BUFFER_SIZE);
 
         AX25_UiFrame_t frame;
         bool ready = false;
